requirements.txt
playwright==1.41.1
pytest==7.4.4
pytest-asyncio==0.23.3
python-dotenv==1.0.0
pymongo==4.6.1
Flask==3.0.0
Flask-PyMongo==2.3.0
Werkzeug==3.0.1
schedule==1.2.1
requests==2.31.0
python-dateutil==2.8.2